+++ src/cards/windows-editor.ts
@@ -1,9 +1,9 @@
 import { customElement } from "lit/decorators.js";
-import type { WindowsConfig } from "./windows";
-import { CarlinkoDeviceEditor } from "./shared-editor";
+import { CarlinkoCabinEditor } from "./cabin-editor";
 
+/** @deprecated Prefer `carlinko-cabin-editor`. */
 @customElement("carlinko-windows-editor")
-export class CarlinkoWindowsEditor extends CarlinkoDeviceEditor<WindowsConfig> {}
+export class CarlinkoWindowsEditor extends CarlinkoCabinEditor {}
 
 declare global {
   interface HTMLElementTagNameMap {